arreglando dise;o y detalles de las blls

+++ Models/PagosDetalle.cs
@@ -6,7 +6,7 @@ public class PagosDetalle
     public int Id { get; set; }
     public int PagoId { get; set; }
     public int PrestamoId { get; set; }
-    public decimal ValorPagado { get; set; }
+    public double ValorPagado { get; set; }
 
   
 }
